+++ vm/wasmvm/util/converts.go
@@ -21,6 +21,7 @@ package util
 import (
 	"encoding/binary"
 	"math"
+	"bytes"
 )
 
 func Float32ToByte(float float32) []byte {
@@ -49,4 +50,30 @@ func ByteToFloat64(bytes []byte) float64 {
 	bits := binary.LittleEndian.Uint64(bytes)
 
 	return math.Float64frombits(bits)
+}
+
+func Int32ToBytes(i32 uint32) []byte {
+	bytesBuffer := bytes.NewBuffer([]byte{})
+	binary.Write(bytesBuffer, binary.LittleEndian, i32)
+	return bytesBuffer.Bytes()
+}
+
+func Int64ToBytes(i64 uint64) []byte {
+	bytesBuffer := bytes.NewBuffer([]byte{})
+	binary.Write(bytesBuffer, binary.LittleEndian, i64)
+	return bytesBuffer.Bytes()
+}
+func Float32ToBytes(float float32) []byte {
+	bits := math.Float32bits(float)
+	bytes := make([]byte, 4)
+	binary.LittleEndian.PutUint32(bytes, bits)
+
+	return bytes
+}
+
+func Float64ToBytes(float float64) []byte {
+	bits := math.Float64bits(float)
+	bytes := make([]byte, 8)
+	binary.LittleEndian.PutUint64(bytes, bits)
+	return bytes
 }
